Update repo layout
