+++ Pipfile
@@ -9,4 +9,4 @@ pylint = "*"
 [packages]
 
 [requires]
-python_version = "3.8"
+python_version = "3"
