+++ README.md
@@ -1,12 +1,27 @@
 # Hash Code Helper
 
-Helper scripts for running hash code problems.
+Helper scripts for running hash code problems
+
+Use this repository to create a new repository for you Hash Code entry. This newly created repository will contain a folder in which to write your submission. Submissions written in the folder can then be run and managed by the `main.py` file that will be located alongside the `submission` folder.
+
+The created repository contains a `main.py` file that allows you to do the folloing:
+- Run your sumbission for a single data file
+- Run your submission for all data files
+- Zip your submission
+- Run for all files and zip your submission 
 
 ## Usage
-* Save the competition data files in the data folder
-* `pipenv install` - Install dependancies
-* `pipenv run python main.py` - Run the helper script
+1. Clone this repository
+1. Run `pipenv run python setup.py <new-repo-location>`
+1. A new repository at `<new-repo-location>` is created. See the README file in this location for usage instructions.
+
+## Tests
+Test for the Hash Code helper are located in the `test` directory. These can be run using the following command:
+
+## Contributing
+All contributions are welcome. Submissions should conform to PEP8 styling standards.
 
 ## Dependancies
 * Python version 3.7.2 or higher
 * pipenv
+* Git
