+++ src/README.template
@@ -1,0 +1,12 @@
+# Hash Code Submission
+
+Google Hash Code Entry. This repo 
+
+## Usage
+* Save the competition data files in the data folder
+* `pipenv install` - Install dependancies
+* `pipenv run python main.py` - Run the helper script
+
+## Dependancies
+* Python version 3.7.2 or higher
+* pipenv
