+++ src/submission/main.py
@@ -1,0 +1,18 @@
+###############################
+##
+# HASH CODE ENTRY
+##
+# Usage: python main.py <path_to_source_file> <output_location>
+##
+###############################
+import sys
+
+
+def main(file_location, output_location):
+    print("Running Hash Code Entry")
+    # Hash Code here :)
+
+
+# When run from the terminal
+if __name__ == '__main__':
+    main(sys.argv[1], sys.argv[2])
